Made adjustments to protocol class so that data is sampled evenly in between updates
Set the sampling rate of the gyro from 16hz to 800 hz. Sampling rate of
gyro is 800hz, and our robot code update is 50hz, so the there should be
800hz/15hz = 16 samples read in between updates to ensure evenly
distributed data samples. Therefore, changed the gyro queue to 16
values. Also reintroduced hardcoded values for time passed for debugging
purposes. (Wow, that was a long commit!)